+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 - Support resource discovery for `waas` resources
+- Support resource discovery for `database` resources: exadata infrastructures, vm clusters, backup destinations, databases, database backups
+
 
 ## 3.79.0 (June 03, 2020)